fix: author profile tests
