--- packages/author-profile/__tests__/author-profile-content.native.test.js
@@ -1,4 +1,0 @@
-import test from "./author-profile-helper";
-import AuthorProfile from "../author-profile.js";
-
-test(AuthorProfile);
