+++ packages/author-profile/package.json
@@ -70,8 +70,8 @@
   "dependencies": {
     "@times-components/author-head": "0.3.1",
     "@times-components/card": "0.4.1",
-    "@times-components/pagination": "0.4.0",
-    "@times-components/provider": "0.4.3",
+    "@times-components/pagination": "0.4.1",
+    "@times-components/provider": "0.4.4",
     "jest-cli": "20.0.4",
     "lodash.get": "4.4.2",
     "lodash.merge": "4.6.0",
